requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
mysql-connector-python==8.0.28
protobuf==3.19.4
requests==2.27.1
schedule==1.1.0
urllib3==1.26.8
